# CLOUD-797 review fix
# CLOUD-803 review fix (crashed in cloud_connect_test_util)
# CLOUD-811 Memory leak in cloud_connectivity_ut
# Disable sending crash reports for developer builds
# Fix stun::AsyncClient asserts and logs
# MOBILE-363 One more fix
# MOBILE-363 Review fix
# Make NX_LOG a function-like call
# QnSettings: add support for config file override like it works for mediaserver
# VMS-4566 fix AsyncHttpClient traffic logs
# VMS-4566 review fix
# fix home dir in bpi core analysis
# introdule nx::network::test::runTest to get rid of ifdef SocketGlobals
# minor wrning fix
# some gcc warning fixes
#CLOUD-801 review fixes
#CLOUD-803 fixed improper async I/O cancellation
#CLOUD-805 fixed compile error
#CLOUD-805 fixed undefined behavior in CloudStreamSocket::shutdown
#CLOUD-807 Added more tests
#CLOUD-807 Added more tests to reproduce the issue
#CLOUD-807 More logs
#CLOUD-813 Coding style fix
#CLOUD-813 Mediator: fixed service destruction order
#CLOUD-813 fixed compile error
#CLOUD-813 rename: StunServerComposite -> StunServer
#CLOUD-818 Fixed: mediator incorrectly parsed cloud_db/endpoint setting
#CLOUD-818 Mediator: added cloud_db/url setting
#UT fixes
#VMS-3357 review fixes
#VMS-4457 Added AbstractPollSet class do decouple aio implementation and actual pollset type. Added wrapper for existing PollSet implementations.
#VMS-4457 Added tests for Pollset class
#VMS-4457 Adding tests
#VMS-4457 Enabled aio.pollsetError test
#VMS-4457 Fixed Pollset tests. Added more
#VMS-4457 Removed files added by mistake
#VMS-4457 Removed old unused source file
#VMS-4457 UDT usage disabled in cloud_db
#VMS-4457 UDT usage disabled in connection_mediator
#VMS-4457 Udt usage is made tunable during network sub-system initialization.
#VMS-5176 api/setupCloudSystem: fixed race condition between saving cloud system credentials and fetching cloud nonce
Backed out changeset: 072a96bf8ba8
Backed out changeset: eb1baee388bf
CLOUD-797 IncomingReverseTunnelConnection minor memory leak
Every caller receives update address completion result
File renamed appropriately
MOBILE-418: Mobile client crashes when trying to open some I/O modules
NX1-30: Tmp as tmpfs on NX1
VMS-2052 Server doesn't guess password root/pass for AxisGet resource properties from the right place.
VMS-2052: Server doesn't guess password root/pass for Axis
VMS-3068: func test
VMS-3069: func test
VMS-3258: RestAPI: Merge incomplete structures with DB contents
VMS-3280: Add 'format=json' in 'Try method' for /ec2/bookmarks/add, /delete,/tags,/updates
VMS-3458: 'Server updated' and 'User updated' show twice
VMS-3692
VMS-4180: func test
VMS-4385: audit trail log length
VMS-4518 Server DB should be always backed up on remove and upgrade
VMS-4599: func test
VMS-4804: API: /api/RecordedTimePeriods works improperly with parameters periodsType and filter
VMS-5011 Do not allow update from version 2.2.1 and less to 3.0.0
VMS-5039 There's no button '2way audio' on Hikvision camera if it was find as 'Unauthorized' first
VMS-5069: Crash detected in client
VMS-5112: ISD 4K camera MJPEG not decoded properly
VMS-5139: Can't open Server logs from Client
VMS-5141: Email Server settings are not saved to Server
VMS-5184: desktop camera found each discovery loop
VMS-5188: Cameras info files UT
VMS-5189: Camera IP conflict has been broken
cloud cases temporary disabled
fixed compile error
fixed compile errors
fixed null functor access
packages.config added to ignore list
perfTest initial
vms-4405. Client CDB result code unit test.
vms-4869. More logs.
vms-5136. Multiserver request params deserializatin fix.
vms-5165. Enable multiple servers running on the same machine.
vms-5187. Unmounted storages filter fixes and UT.
vms. Storage plugin sample fix.

<==== Mercurial data ====>
Branch: prod_3.0.0

+++ client/libclient/src/ui/widgets/properties/storage_analytics_widget.cpp
@@ -448,8 +448,8 @@ qint64 QnStorageAnalyticsWidget::sliderPositionToBytes(int value) const
         return 0;
 
     int idx = value / kTicksPerInterval;
-    if (idx >= kExtraDataBase.size() - 1)
-        return kExtraDataBase.back();
+    if (idx >= (qint64) kExtraDataBase.size() - 1)
+        return (qint64) kExtraDataBase.back();
 
     qint64 k1 = kExtraDataBase[idx];
     qint64 k2 = kExtraDataBase[idx+1];
@@ -462,7 +462,7 @@ qint64 QnStorageAnalyticsWidget::sliderPositionToBytes(int value) const
 int QnStorageAnalyticsWidget::bytesToSliderPosition (qint64 value) const
 {
     int idx = 0;
-    for (; idx < kExtraDataBase.size() - 1; ++idx)
+    for (; idx < (qint64) kExtraDataBase.size() - 1; ++idx)
     {
         if (kExtraDataBase[idx+1] >= value)
             break;
